requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
